backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
python-jose[cryptography]
python-multipart
huggingface_hub
pytest
